refactor(tangle-dapp): Move Polkadot API ctx to shared ui
