+++ apps/tangle-dapp/app/blueprints/[id]/useBlueprintDetails.ts
@@ -7,6 +7,7 @@ import {
   TangleErrorCode,
 } from '@webb-tools/tangle-shared-ui/types/error';
 import type { OperatorMap } from '@webb-tools/tangle-shared-ui/types/restake';
+import type { AssetMap } from '@webb-tools/tangle-shared-ui/types/restake';
 import { useCallback } from 'react';
 import { combineLatest, switchMap } from 'rxjs';
 
@@ -17,7 +18,6 @@ import useRestakeDelegatorInfo from '../../../data/restake/useRestakeDelegatorIn
 import useRestakeOperatorMap from '../../../data/restake/useRestakeOperatorMap';
 import useRestakeTVL from '../../../data/restake/useRestakeTVL';
 import type { Blueprint, Operator } from '../../../types/blueprint';
-import type { AssetMap } from '../../../types/restake';
 import {
   getAccountInfo,
   getMultipleAccountInfo,
