+++ apps/tangle-dapp/app/blueprints/layout.tsx
@@ -1,6 +1,6 @@
+import { PolkadotApiProvider } from '@webb-tools/tangle-shared-ui/context/PolkadotApiContext';
 import { PropsWithChildren } from 'react';
 
-import { PolkadotApiProvider } from '../../context/PolkadotApiContext';
 import { RestakeContextProvider } from '../../context/RestakeContext';
 
 export default function Layout({ children }: PropsWithChildren) {
