+++ apps/tangle-dapp/app/restake/providers.tsx
@@ -1,9 +1,9 @@
 'use client';
 
+import { PolkadotApiProvider } from '@webb-tools/tangle-shared-ui/context/PolkadotApiContext';
 import { NuqsAdapter } from 'nuqs/adapters/next/app';
 import { type PropsWithChildren } from 'react';
 
-import { PolkadotApiProvider } from '../../context/PolkadotApiContext';
 import { RestakeContextProvider } from '../../context/RestakeContext';
 
 export default function Providers({ children }: PropsWithChildren) {
