+++ apps/tangle-dapp/app/restake/utils.ts
@@ -1,7 +1,7 @@
 import type { OperatorDelegatorBond } from '@webb-tools/tangle-shared-ui/types/restake';
+import type { AssetMap } from '@webb-tools/tangle-shared-ui/types/restake';
 
 import type { VaultToken } from '../../components/tables/Operators/types';
-import type { AssetMap } from '../../types/restake';
 import safeFormatUnits from '../../utils/safeFormatUnits';
 
 export function calculateTimeRemaining(
