+++ apps/tangle-dapp/data/restake/useDelegatorTVL.ts
@@ -1,7 +1,8 @@
+import type { AssetMap } from '@webb-tools/tangle-shared-ui/types/restake';
 import { useObservable, useObservableState } from 'observable-hooks';
 import { of, switchMap } from 'rxjs';
 
-import type { AssetMap, DelegatorInfo } from '../../types/restake';
+import type { DelegatorInfo } from '../../types/restake';
 import safeFormatUnits from '../../utils/safeFormatUnits';
 
 export function useDelegatorTVL(
