+++ apps/tangle-dapp/data/restake/useRestakeDelegatorInfo.ts
@@ -5,10 +5,10 @@ import type {
   PalletMultiAssetDelegationDelegatorWithdrawRequest,
 } from '@polkadot/types/lookup';
 import { WebbError, WebbErrorCodes } from '@webb-tools/dapp-types/WebbError';
+import usePolkadotApi from '@webb-tools/tangle-shared-ui/hooks/usePolkadotApi';
 import { useObservable, useObservableState } from 'observable-hooks';
 import { map, of, switchMap } from 'rxjs';
 
-import usePolkadotApi from '../../hooks/usePolkadotApi';
 import useSubstrateAddress from '../../hooks/useSubstrateAddress';
 import type { DelegatorInfo } from '../../types/restake';
