+++ apps/tangle-dapp/data/restake/useRestakeOperatorMap.ts
@@ -5,6 +5,7 @@ import type {
   PalletMultiAssetDelegationOperatorOperatorStatus,
 } from '@polkadot/types/lookup';
 import isDefined from '@webb-tools/dapp-types/utils/isDefined';
+import usePolkadotApi from '@webb-tools/tangle-shared-ui/hooks/usePolkadotApi';
 import type {
   OperatorMap,
   OperatorMetadata,
@@ -12,8 +13,6 @@ import type {
 import { useObservableState } from 'observable-hooks';
 import { useMemo } from 'react';
 import { map, type Observable, of } from 'rxjs';
-
-import usePolkadotApi from '../../hooks/usePolkadotApi';
 
 type UseRestakeOperatorMapReturnType = {
   operatorMap: OperatorMap;
