+++ libs/tangle-shared-ui/src/context/PolkadotApiContext.tsx
@@ -4,11 +4,6 @@ import { WsProvider } from '@polkadot/api';
 import { ApiPromise } from '@polkadot/api/promise';
 import { ApiRx } from '@polkadot/api/rx';
 import type { Maybe } from '@webb-tools/dapp-types/utils/types';
-import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
-import {
-  getApiPromise,
-  getApiRx,
-} from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
 import noop from 'lodash/noop';
 import {
   createContext,
@@ -21,6 +16,8 @@ import {
 } from 'react';
 import useSWRImmutable from 'swr/immutable';
 import { Prettify } from 'viem/chains';
+import useNetworkStore from './useNetworkStore';
+import { getApiPromise, getApiRx } from '../utils/polkadot/api';
 
 export type PolkadotApiContextProps = Prettify<{
   apiPromise: ApiPromise;
